+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 
-## v0.12.x
+## v1.2.1
 
 ### 💥️ Breaking Change